requirements.txt
flask>=2.0.1
tensorflow>=2.6.0
pillow>=8.3.1
numpy>=1.19.5
matplotlib>=3.4.3
pandas>=1.3.3
seaborn>=0.11.2
